sh format: Separate inline extents messsage
Building on the change made by 4d28a4d58f17af32a3be9a12736404bf25003918:

Printing "already reflinked to original" in this case is confusing,
especially because a) that's not true and b) that message should never
appear in --keep-reflinked mode as the files will be marked as originals
in advance. Also, apply this logic to cp --reflink as well.