(#614) shredder: only hash single-inode groups that are actually hardlinked
`group->n_inodes == 1` is also true for groups that simply consist of a
single file. This condition will cause all single-file groups to be
hashed if `--merge-directories` is also set.

Additionally, the whole `group->n_inodes == 1` condition is redundant
because not following on the branch means that `group->n_clusters == 1`
and therefore `group->n_inodes == 1`.

Thus, check that we are actually dealing with a group of hardlinks (and
not just a single-file group that did not cause an early return because
we are also doing `--hash-unmatched`).

Fixes #614.