(#614) shredder: fix --hash-unmatched criterion
The intention behind this criterion seems to handle partially-hashed
(i. e. child) groups that do not fulfill requirements to continue
hashing them (i. e. only contain a single file after sifting). Normally
such groups are left alone (because we have hashed enough of the file to
understand it is unique), however by definition of --hash-unmatched we
have to hash these files to the end.

The problem is that when rm_shred_group_update_status() is called,
group->held_files does not contain the currently pushed file. Thus,
because this condition is only executed for single-file groups
(group->n_clusters == 1), group->held_files will never be non-empty.
Anyway, even if it was non-empty, head->digest will always be NULL
because held files have digests freed in rm_shred_group_push_file().

Therefore, this condition will never be fulfilled. The only reason why
--hash-unmatched even works at all is due to the broken hardlinked files
condition that matches all single-inode groups. (Another consequence is
that --hash-unmatched is broken if --merge-directories is not used.)