sh: cp_reflink: do not delete the hierarchy, preserve original xattrs
Due to the fact that we cannot preserve xattrs in the stamp directory
(because it is likely on /tmpfs) we are reusing the originals' xattrs,
thus desynchronizing them with files' mtimes. This breaks rmlint's own
--{read,write}-xattrs.

Thus, play a trick and do not delete the duplicate directory at all.
At the same time, do not even bother to copy the originals' attributes
because they are going to be overwritten anyway.